refactor: Separate file for template creation modal
